Complete local AI pipeline with semantic search
This milestone completes Memoria's local AI backend.

Major additions:
- Ollama embedding generation
- Local AI summarization
- PostgreSQL pgvector integration
- Hibernate 7 native vector mapping
- Hybrid semantic search
- Asynchronous summary generation
- Event-driven AI pipeline

Major issue resolved:

The largest blocker was PostgreSQL vector retrieval.

Initially, embeddings were generated successfully and persisted into PostgreSQL, but Hibernate failed during entity retrieval with:

org.postgresql.util.PSQLException:
No results were returned by the query.

Root cause:
Hibernate attempted to deserialize PostgreSQL's vector type as a SQL array, while pgvector exposes it as a custom PostgreSQL type.

Solution:
Migrated to Hibernate 7 native vector support using hibernate-vector, SqlTypes.VECTOR, and float[] mapping.

Other changes:
- Replaced Anthropic summarization with Ollama.
- Preserved asynchronous event-driven architecture.
- Improved resilience when local AI services are unavailable.
- Cleaned repository and prepared the backend for release.

Current status:
Stable backend milestone.
Ready for frontend development and AI expansion.

+++ src/main/java/com/memoria/Memoria/repositories/NoteRepository.java
@@ -63,5 +63,5 @@ public interface NoteRepository extends JpaRepository<Note, Long> {
             """, nativeQuery = true)
     List<Object[]> findHybridSearchResults(@Param("userId") Long userId, 
                                           @Param("query") String query, 
-                                          @Param("queryVector") double[] queryVector);
+                                          @Param("queryVector") float[] queryVector);
 }
