+++ src/main/java/com/memoria/Memoria/services/EmbeddingService.java
@@ -3,5 +3,5 @@ package com.memoria.Memoria.services;
 import reactor.core.publisher.Mono;
 
 public interface EmbeddingService {
-    Mono<double[]> getEmbedding(String text);
+    Mono<float[]> getEmbedding(String text);
 }
